+++ automation/jira_qa/__init__.py
@@ -1,0 +1,1 @@
+"""Local Jira QA automation for ClawIQ."""